cleaned up the frontend
