+++ minecraft-platform-frontend/src/App.tsx
@@ -4,7 +4,7 @@ import './App.css';
 import axios from 'axios';
 // import material UI theme
 import { ThemeProvider, createTheme } from '@mui/material/styles';
-import Home from './components/Home';
+import { Home } from './components/Home';
 // auth with aws cognito from amazon-cognito-identity-js
 import { config } from 'process';
 
@@ -16,7 +16,7 @@ function App() {
   return (
     <ThemeProvider theme={theme}>
       <div className="App" >
-        <Home />
+        <Home debug />
       </div >
     </ThemeProvider>
   );
