+++ minecraft-platform-frontend/src/components/Home/index.tsx
@@ -1,0 +1,5 @@
+import Home from "./Home"
+
+export {
+    Home
+}
